test(*): Car racing discrete test passed.
Early terminate the game at 1000 steps. Get 600 - 800 scores in 600K training steps.

TODO: Try early terminate depends on past reward records, because when the car hits a sharp curve and out of the track, it is very hard for the car to come back the track, after that point, continue the game is meaningless.